# openapi.yaml
openapi: 3.0.3
info:
  title: TalkBlog Back API
  version: 1.0.0
  description: OpenAPI spec for TalkBlog backend services
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: User register
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Face login
      operationId: faceLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaceLoginRequest'
      responses:
        '200':
          description: Login result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/auth/password-login:
    post:
      tags:
        - Auth
      summary: Password login
      operationId: passwordLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordLoginRequest'
      responses:
        '200':
          description: Login result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/comments:
    post:
      tags:
        - Comments
      summary: Create a comment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
          multipart/form-data:
            schema:
              type: object
              required:
                - blogId
                - content
              properties:
                blogId:
                  type: integer
                  format: int64
                content:
                  type: string
                  description: 评论内容，支持 Emoji，按 Unicode 码点计数
                parentId:
                  type: integer
                  format: int64
                replyToUserId:
                  type: integer
                  format: int64
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - Comments
      summary: Get comments for a blog
      parameters:
        - name: blogId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: parentId
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Comment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/comments/{commentId}/like:
    post:
      tags:
        - Comments
      summary: Like/unlike a comment
      security:
        - BearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Like result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/comments/{commentId}:
    delete:
      tags:
        - Comments
      summary: Delete own comment
      security:
        - BearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Comment deleted
  /api/comments/mine:
    get:
      tags:
        - Comments
      summary: Get my comments
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: My comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/admin/comments:
    get:
      tags:
        - Admin Comments
      summary: Admin paginated comment list
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: integer
        - name: blogId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Comment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/admin/comments/{commentId}/status:
    patch:
      tags:
        - Admin Comments
      summary: Review comment (approve/reject)
      security:
        - BearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentStatusRequest'
      responses:
        '200':
          description: Review result
  /api/admin/comments/{commentId}:
    delete:
      tags:
        - Admin Comments
      summary: Admin delete comment
      security:
        - BearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Comment deleted
  /api/admin/comments/batch-status:
    patch:
      tags:
        - Admin Comments
      summary: Batch review comments
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchReviewRequest'
      responses:
        '200':
          description: Batch review result
  /api/admin/comments/stats:
    get:
      tags:
        - Admin Comments
      summary: Comment statistics
      security:
        - BearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Statistics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - username
        - password
        - image
      properties:
        email:
          type: string
        username:
          type: string
        password:
          type: string
        image:
          type: string
          description: Base64 face image
    FaceLoginRequest:
      type: object
      required:
        - email
        - image
      properties:
        email:
          type: string
        image:
          type: string
          description: Base64 face image
    PasswordLoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
    CreateCommentRequest:
      type: object
      required:
        - blogId
        - content
      properties:
        blogId:
          type: integer
          format: int64
        content:
          type: string
          maxLength: 500
          description: 评论内容，支持 Emoji，按实际 Unicode 码点计数（最多 500 个码点）
        parentId:
          type: integer
          format: int64
        replyToUserId:
          type: integer
          format: int64
    UpdateCommentStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: integer
          description: "0=pending, 1=approved, 2=rejected"
    BatchReviewRequest:
      type: object
      required:
        - commentIds
        - status
      properties:
        commentIds:
          type: array
          items:
            type: integer
            format: int64
        status:
          type: integer
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
...